+++ CHANGELOG.md
@@ -489,6 +489,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 - Recursion
 - Orientation
 
+## 0104 - Maximum Depth of Binary Tree
+
+- Recursion
+
 ## 0415 - Add strings
 
 - Full adder